fix: Enhance OPC UA node browsing and connection handling
